Make ysprod181003550 fits explicit as lambdas
Dependancy on a pickle file does not work well accorss Python versions.
Instread explicit lambda function representations of the fits for
ysprod181003550 replace the fits loaded from the pickle dump. This is
similar to how cw181003550 is implemented.

With 16 digiits of decimal place accuracy (allowing for scientific
notation) the evaluated fit with explicit lambdas agrees with the pickle
dump implementation to greated than 10 significant figures.